Separate Capture, Pwm and Qei traits into modules
