+++ src/lib.rs
@@ -697,295 +697,11 @@ pub mod spi;
 pub mod timer;
 pub mod watchdog;
 
-/// Input capture
-///
-/// # Examples
-///
-/// You can use this interface to measure the period of (quasi) periodic signals
-/// / events
-///
-/// ```
-/// extern crate embedded_hal as hal;
-/// #[macro_use(block)]
-/// extern crate nb;
-///
-/// use hal::prelude::*;
-///
-/// fn main() {
-///     let mut capture: Capture1 = {
-///         // ..
-/// #       Capture1
-///     };
-///
-///     capture.try_set_resolution(1.ms()).unwrap();
-///
-///     let before = block!(capture.try_capture(Channel::_1)).unwrap();
-///     let after = block!(capture.try_capture(Channel::_1)).unwrap();
-///
-///     let period = after.wrapping_sub(before);
-///
-///     println!("Period: {} ms", period);
-/// }
-///
-/// # use core::convert::Infallible;
-/// # struct MilliSeconds(u32);
-/// # trait U32Ext { fn ms(self) -> MilliSeconds; }
-/// # impl U32Ext for u32 { fn ms(self) -> MilliSeconds { MilliSeconds(self) } }
-/// # struct Capture1;
-/// # enum Channel { _1 }
-/// # impl hal::Capture for Capture1 {
-/// #     type Error = Infallible;
-/// #     type Capture = u16;
-/// #     type Channel = Channel;
-/// #     type Time = MilliSeconds;
-/// #     fn try_capture(&mut self, _: Channel) -> ::nb::Result<u16, Self::Error> { Ok(0) }
-/// #     fn try_disable(&mut self, _: Channel) -> Result<(), Self::Error> { unimplemented!() }
-/// #     fn try_enable(&mut self, _: Channel) -> Result<(), Self::Error> { unimplemented!() }
-/// #     fn try_get_resolution(&self) -> Result<MilliSeconds, Self::Error> { unimplemented!() }
-/// #     fn try_set_resolution<T>(&mut self, _: T) -> Result<(), Self::Error> where T: Into<MilliSeconds> { Ok(()) }
-/// # }
-/// ```
-// unproven reason: pre-singletons API. With singletons a `CapturePin` (cf. `PwmPin`) trait seems more
-// appropriate
-pub trait Capture {
-    /// Enumeration of `Capture` errors
-    ///
-    /// Possible errors:
-    ///
-    /// - *overcapture*, the previous capture value was overwritten because it
-    ///   was not read in a timely manner
-    type Error;
+mod capture;
+pub use capture::Capture;
 
-    /// Enumeration of channels that can be used with this `Capture` interface
-    ///
-    /// If your `Capture` interface has no channels you can use the type `()`
-    /// here
-    type Channel;
+mod pwm;
+pub use pwm::{Pwm, PwmPin};
 
-    /// A time unit that can be converted into a human time unit (e.g. seconds)
-    type Time;
-
-    /// The type of the value returned by `capture`
-    type Capture;
-
-    /// "Waits" for a transition in the capture `channel` and returns the value
-    /// of counter at that instant
-    ///
-    /// NOTE that you must multiply the returned value by the *resolution* of
-    /// this `Capture` interface to get a human time unit (e.g. seconds)
-    fn try_capture(&mut self, channel: Self::Channel) -> nb::Result<Self::Capture, Self::Error>;
-
-    /// Disables a capture `channel`
-    fn try_disable(&mut self, channel: Self::Channel) -> Result<(), Self::Error>;
-
-    /// Enables a capture `channel`
-    fn try_enable(&mut self, channel: Self::Channel) -> Result<(), Self::Error>;
-
-    /// Returns the current resolution
-    fn try_get_resolution(&self) -> Result<Self::Time, Self::Error>;
-
-    /// Sets the resolution of the capture timer
-    fn try_set_resolution<R>(&mut self, resolution: R) -> Result<(), Self::Error>
-    where
-        R: Into<Self::Time>;
-}
-
-/// Pulse Width Modulation
-///
-/// # Examples
-///
-/// Use this interface to control the power output of some actuator
-///
-/// ```
-/// extern crate embedded_hal as hal;
-///
-/// use hal::prelude::*;
-///
-/// fn main() {
-///     let mut pwm: Pwm1 = {
-///         // ..
-/// #       Pwm1
-///     };
-///
-///     pwm.try_set_period(1.khz()).unwrap();
-///
-///     let max_duty = pwm.try_get_max_duty().unwrap();
-///
-///     // brightest LED
-///     pwm.try_set_duty(Channel::_1, max_duty).unwrap();
-///
-///     // dimmer LED
-///     pwm.try_set_duty(Channel::_2, max_duty / 4).unwrap();
-/// }
-///
-/// # use core::convert::Infallible;
-/// # struct KiloHertz(u32);
-/// # trait U32Ext { fn khz(self) -> KiloHertz; }
-/// # impl U32Ext for u32 { fn khz(self) -> KiloHertz { KiloHertz(self) } }
-/// # enum Channel { _1, _2 }
-/// # struct Pwm1;
-/// # impl hal::Pwm for Pwm1 {
-/// #     type Error = Infallible;
-/// #     type Channel = Channel;
-/// #     type Time = KiloHertz;
-/// #     type Duty = u16;
-/// #     fn try_disable(&mut self, _: Channel) -> Result<(), Self::Error> { unimplemented!() }
-/// #     fn try_enable(&mut self, _: Channel) -> Result<(), Self::Error> { unimplemented!() }
-/// #     fn try_get_duty(&self, _: Channel) -> Result<u16, Self::Error> { unimplemented!() }
-/// #     fn try_get_max_duty(&self) -> Result<u16, Self::Error> { Ok(0) }
-/// #     fn try_set_duty(&mut self, _: Channel, _: u16) -> Result<(), Self::Error> { Ok(()) }
-/// #     fn try_get_period(&self) -> Result<KiloHertz, Self::Error> { unimplemented!() }
-/// #     fn try_set_period<T>(&mut self, _: T) -> Result<(), Self::Error> where T: Into<KiloHertz> { Ok(()) }
-/// # }
-/// ```
-// unproven reason: pre-singletons API. The `PwmPin` trait seems more useful because it models independent
-// PWM channels. Here a certain number of channels are multiplexed in a single implementer.
-pub trait Pwm {
-    /// Enumeration of `Pwm` errors
-    type Error;
-
-    /// Enumeration of channels that can be used with this `Pwm` interface
-    ///
-    /// If your `Pwm` interface has no channels you can use the type `()`
-    /// here
-    type Channel;
-
-    /// A time unit that can be converted into a human time unit (e.g. seconds)
-    type Time;
-
-    /// Type for the `duty` methods
-    ///
-    /// The implementer is free to choose a float / percentage representation
-    /// (e.g. `0.0 .. 1.0`) or an integer representation (e.g. `0 .. 65535`)
-    type Duty;
-
-    /// Disables a PWM `channel`
-    fn try_disable(&mut self, channel: Self::Channel) -> Result<(), Self::Error>;
-
-    /// Enables a PWM `channel`
-    fn try_enable(&mut self, channel: Self::Channel) -> Result<(), Self::Error>;
-
-    /// Returns the current PWM period
-    fn try_get_period(&self) -> Result<Self::Time, Self::Error>;
-
-    /// Returns the current duty cycle
-    fn try_get_duty(&self, channel: Self::Channel) -> Result<Self::Duty, Self::Error>;
-
-    /// Returns the maximum duty cycle value
-    fn try_get_max_duty(&self) -> Result<Self::Duty, Self::Error>;
-
-    /// Sets a new duty cycle
-    fn try_set_duty(&mut self, channel: Self::Channel, duty: Self::Duty)
-        -> Result<(), Self::Error>;
-
-    /// Sets a new PWM period
-    fn try_set_period<P>(&mut self, period: P) -> Result<(), Self::Error>
-    where
-        P: Into<Self::Time>;
-}
-
-/// A single PWM channel / pin
-///
-/// See `Pwm` for details
-pub trait PwmPin {
-    /// Enumeration of `PwmPin` errors
-    type Error;
-
-    /// Type for the `duty` methods
-    ///
-    /// The implementer is free to choose a float / percentage representation
-    /// (e.g. `0.0 .. 1.0`) or an integer representation (e.g. `0 .. 65535`)
-    type Duty;
-
-    /// Disables a PWM `channel`
-    fn try_disable(&mut self) -> Result<(), Self::Error>;
-
-    /// Enables a PWM `channel`
-    fn try_enable(&mut self) -> Result<(), Self::Error>;
-
-    /// Returns the current duty cycle
-    fn try_get_duty(&self) -> Result<Self::Duty, Self::Error>;
-
-    /// Returns the maximum duty cycle value
-    fn try_get_max_duty(&self) -> Result<Self::Duty, Self::Error>;
-
-    /// Sets a new duty cycle
-    fn try_set_duty(&mut self, duty: Self::Duty) -> Result<(), Self::Error>;
-}
-
-/// Quadrature encoder interface
-///
-/// # Examples
-///
-/// You can use this interface to measure the speed of a motor
-///
-/// ```
-/// extern crate embedded_hal as hal;
-/// #[macro_use(block)]
-/// extern crate nb;
-///
-/// use hal::prelude::*;
-///
-/// fn main() {
-///     let mut qei: Qei1 = {
-///         // ..
-/// #       Qei1
-///     };
-///     let mut timer: Timer6 = {
-///         // ..
-/// #       Timer6
-///     };
-///
-///
-///     let before = qei.try_count().unwrap();
-///     timer.try_start(1.s()).unwrap();
-///     block!(timer.try_wait());
-///     let after = qei.try_count().unwrap();
-///
-///     let speed = after.wrapping_sub(before);
-///     println!("Speed: {} pulses per second", speed);
-/// }
-///
-/// # use core::convert::Infallible;
-/// # struct Seconds(u32);
-/// # trait U32Ext { fn s(self) -> Seconds; }
-/// # impl U32Ext for u32 { fn s(self) -> Seconds { Seconds(self) } }
-/// # struct Qei1;
-/// # impl hal::Qei for Qei1 {
-/// #     type Error = Infallible;
-/// #     type Count = u16;
-/// #     fn try_count(&self) -> Result<u16, Self::Error> { Ok(0) }
-/// #     fn try_direction(&self) -> Result<::hal::Direction, Self::Error> { unimplemented!() }
-/// # }
-/// # struct Timer6;
-/// # impl hal::timer::CountDown for Timer6 {
-/// #     type Error = Infallible;
-/// #     type Time = Seconds;
-/// #     fn try_start<T>(&mut self, _: T) -> Result<(), Infallible> where T: Into<Seconds> { Ok(()) }
-/// #     fn try_wait(&mut self) -> ::nb::Result<(), Infallible> { Ok(()) }
-/// # }
-/// ```
-// unproven reason: needs to be re-evaluated in the new singletons world. At the very least this needs a
-// reference implementation
-pub trait Qei {
-    /// Enumeration of `Qei` errors
-    type Error;
-
-    /// The type of the value returned by `count`
-    type Count;
-
-    /// Returns the current pulse count of the encoder
-    fn try_count(&self) -> Result<Self::Count, Self::Error>;
-
-    /// Returns the count direction
-    fn try_direction(&self) -> Result<Direction, Self::Error>;
-}
-
-/// Count direction
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum Direction {
-    /// 3, 2, 1
-    Downcounting,
-    /// 1, 2, 3
-    Upcounting,
-}
+mod qei;
+pub use qei::{Direction, Qei};
